requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
openai>=1.51.0
pyyaml>=6.0.0
python-multipart>=0.0.9

